requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
DateTime==4.3
discord.py==1.4.1
docopt==0.6.2
idna==2.10
idna-ssl==1.1.0
monotonic==1.5
multidict==4.7.6
numpy==1.19.2
peewee==3.13.3
peewee-migrations==0.3.19
pycryptodomex==3.9.8
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
six==1.15.0
steam==1.0.2
typing-extensions==3.7.4.3
urllib3==1.25.10
vdf==3.3
yarl==1.5.1
zope.interface==5.1.0
